+++ packages/twenty-server/src/engine/twenty-orm/pg-shared-pool/pg-shared-pool.module.ts
@@ -1,13 +1,12 @@
 import {
-    Global,
-    Module,
-    OnApplicationShutdown,
-    OnModuleInit,
+  Global,
+  Module,
+  OnApplicationShutdown,
+  OnModuleInit,
 } from '@nestjs/common';
 
 import { TwentyConfigModule } from 'src/engine/core-modules/twenty-config/twenty-config.module';
-import { TwentyConfigService } from 'src/engine/core-modules/twenty-config/twenty-config.service';
-import { PgPoolSharedService } from 'src/engine/twenty-orm/services/pg-shared-pool.service';
+import { PgPoolSharedService } from 'src/engine/twenty-orm/pg-shared-pool/pg-shared-pool.service';
 
 /**
  * Module that initializes the shared pg pool at application bootstrap
@@ -19,10 +18,7 @@ import { PgPoolSharedService } from 'src/engine/twenty-orm/services/pg-shared-po
   exports: [PgPoolSharedService],
 })
 export class PgPoolSharedModule implements OnModuleInit, OnApplicationShutdown {
-  constructor(
-    private readonly pgPoolSharedService: PgPoolSharedService,
-    private readonly configService: TwentyConfigService,
-  ) {}
+  constructor(private readonly pgPoolSharedService: PgPoolSharedService) {}
 
   /**
    * Initialize the pool sharing service when the module is initialized
